requirements.txt
apache-airflow>=2.7.0
pandas>=1.5.0
python-dotenv>=1.0.0
requests>=2.31.0
databricks-sql-connector>=3.0.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
apache-airflow-providers-google>=10.9.0
